API: Rename symbols directory to vacuum
