+++ docs/source/_static/icons/ApertureValveSymbolIcon.png
