+++ docs/source/_static/icons/CathodeGaugeSymbolIcon.png
