+++ docs/source/_static/icons/ColdCathodeGaugeSymbolIcon.png
