+++ docs/source/_static/icons/FastShutterSymbolIcon.png
