+++ docs/source/_static/icons/GetterPumpSymbolIcon.png
