+++ docs/source/_static/icons/HotCathodeGaugeSymbolIcon.png
