+++ docs/source/_static/icons/IonPumpSymbolIcon.png
