+++ docs/source/_static/icons/NeedleValveSymbolIcon.png
