+++ docs/source/_static/icons/PiraniGaugeSymbolIcon.png
