+++ docs/source/_static/icons/PneumaticValveSymbolIcon.png
