+++ docs/source/_static/icons/ProportionalValveSymbolIcon.png
