+++ docs/source/_static/icons/RGASymbolIcon.png
