+++ docs/source/_static/icons/RightAngleManualValveSymbolIcon.png
