+++ docs/source/_static/icons/ScrollPumpSymbolIcon.png
