+++ docs/source/_static/icons/TurboPumpSymbolIcon.png
