--- docs/source/symbols/gauges.rst
@@ -1,21 +1,0 @@
-========
-Gauges
-========
-
-Pirani Gauge
----------------
-
-.. autoclass:: pcdswidgets.symbols.gauges.PiraniGauge
-   :members:
-
-Hot Cathode Gauge
------------------
-
-.. autoclass:: pcdswidgets.symbols.gauges.HotCathodeGauge
-   :members:
-
-Cold Cathode Gauge
-------------------
-
-.. autoclass:: pcdswidgets.symbols.gauges.ColdCathodeGauge
-   :members:
