--- docs/source/symbols/pumps.rst
@@ -1,27 +1,0 @@
-========
-Pumps
-========
-
-Ion Pump
---------
-
-.. autoclass:: pcdswidgets.symbols.pumps.IonPump
-   :members:
-
-Turbo Pump
-----------
-
-.. autoclass:: pcdswidgets.symbols.pumps.TurboPump
-   :members:
-
-Scroll Pump
------------
-
-.. autoclass:: pcdswidgets.symbols.pumps.ScrollPump
-   :members:
-
-Getter Pump
------------
-
-.. autoclass:: pcdswidgets.symbols.pumps.GetterPump
-   :members:
