--- docs/source/symbols/valves.rst
@@ -1,39 +1,0 @@
-========
-Valves
-========
-
-Pneumatic Valve
----------------
-
-.. autoclass:: pcdswidgets.symbols.valves.PneumaticValve
-   :members:
-
-Aperture Valve
---------------
-
-.. autoclass:: pcdswidgets.symbols.valves.ApertureValve
-   :members:
-
-Fast Shutter
-------------
-
-.. autoclass:: pcdswidgets.symbols.valves.FastShutter
-   :members:
-
-Needle Valve
-------------
-
-.. autoclass:: pcdswidgets.symbols.valves.NeedleValve
-   :members:
-
-Proportional Valve
-------------------
-
-.. autoclass:: pcdswidgets.symbols.valves.ProportionalValve
-   :members:
-
-Right Angle Manual Valve
-------------------------
-
-.. autoclass:: pcdswidgets.symbols.valves.RightAngleManualValve
-   :members:
