+++ docs/source/vacuum/gauges.rst
@@ -1,0 +1,21 @@
+========
+Gauges
+========
+
+Pirani Gauge
+---------------
+
+.. autoclass:: pcdswidgets.vacuum.gauges.PiraniGauge
+   :members:
+
+Hot Cathode Gauge
+-----------------
+
+.. autoclass:: pcdswidgets.vacuum.gauges.HotCathodeGauge
+   :members:
+
+Cold Cathode Gauge
+------------------
+
+.. autoclass:: pcdswidgets.vacuum.gauges.ColdCathodeGauge
+   :members:
