+++ docs/source/vacuum/valves.rst
@@ -1,0 +1,39 @@
+========
+Valves
+========
+
+Pneumatic Valve
+---------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.PneumaticValve
+   :members:
+
+Aperture Valve
+--------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.ApertureValve
+   :members:
+
+Fast Shutter
+------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.FastShutter
+   :members:
+
+Needle Valve
+------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.NeedleValve
+   :members:
+
+Proportional Valve
+------------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.ProportionalValve
+   :members:
+
+Right Angle Manual Valve
+------------------------
+
+.. autoclass:: pcdswidgets.vacuum.valves.RightAngleManualValve
+   :members:
