+++ pcdswidgets/vacuum/gauges.py
